Move pytest entrypoints to subdirectory that does not import ddtrace
