+++ setup.cfg
@@ -11,6 +11,7 @@ ignore-words-list = asend,dne,fo,medias,ment,nin,ot,setttings,statics,ba,spawnve
 # --cov-report is intentionally empty else pytest-cov will default to generating a report
 addopts =
   --cov=ddtrace/
+  --cov=_ddtrace/
   --cov=tests/
   --cov-append
   --cov-report=
